Project structure changed: individual_subgroups renamed to subgroup_sets ; all algoritms (nominal and numeric) moved to the same folder.
