+++ subgroups/algorithms/__init__.py
@@ -1,11 +1,11 @@
 from subgroups.algorithms.algorithm import Algorithm
-from subgroups.algorithms.individual_subgroups.nominal_target.sdmap import SDMap
-from subgroups.algorithms.individual_subgroups.nominal_target.vlsd import VLSD
-from subgroups.algorithms.individual_subgroups.nominal_target.bsd import BSD
-from subgroups.algorithms.individual_subgroups.nominal_target.cbsd import CBSD
-from subgroups.algorithms.individual_subgroups.nominal_target.cpbsd import CPBSD
-from subgroups.algorithms.individual_subgroups.nominal_target.sdmapstar import SDMapStar
-from subgroups.algorithms.individual_subgroups.nominal_target.qfinder import QFinder
-from subgroups.algorithms.subgroup_lists.nominal_target.gmsl import GMSL
-from subgroups.algorithms.subgroup_lists.nominal_target.psld import PSLD
-from subgroups.algorithms.subgroup_lists.nominal_target.dslm import DSLM
+from subgroups.algorithms.subgroup_sets.sdmap import SDMap
+from subgroups.algorithms.subgroup_sets.vlsd import VLSD
+from subgroups.algorithms.subgroup_sets.bsd import BSD
+from subgroups.algorithms.subgroup_sets.cbsd import CBSD
+from subgroups.algorithms.subgroup_sets.cpbsd import CPBSD
+from subgroups.algorithms.subgroup_sets.sdmapstar import SDMapStar
+from subgroups.algorithms.subgroup_sets.qfinder import QFinder
+from subgroups.algorithms.subgroup_lists.gmsl import GMSL
+from subgroups.algorithms.subgroup_lists.psld import PSLD
+from subgroups.algorithms.subgroup_lists.dslm import DSLM
