+++ subgroups/tests/algorithms/subgroup_sets/test_bsd.py
@@ -9,7 +9,7 @@
 from os import remove
 from bitarray import bitarray
 from pandas import DataFrame
-from subgroups.algorithms.individual_subgroups.nominal_target.bsd import BSD
+from subgroups.algorithms.subgroup_sets.bsd import BSD
 from subgroups.core.operator import Operator
 from subgroups.core.pattern import Pattern
 from subgroups.core.selector import Selector
