+++ subgroups/tests/algorithms/subgroup_sets/test_qfinder.py
@@ -9,7 +9,7 @@
 from os import remove
 from bitarray import bitarray
 from pandas import DataFrame
-from subgroups.algorithms.individual_subgroups.nominal_target.qfinder import QFinder
+from subgroups.algorithms.subgroup_sets.qfinder import QFinder
 from subgroups.core.operator import Operator
 from subgroups.core.pattern import Pattern
 from subgroups.core.selector import Selector
